CommunityMonitor Modal file rename
