+++ front/app/containers/App/ModalQueue/modals/CommunityMonitor/Modal.tsx
